requirements.txt
click>=8.1
jinja2>=3.1
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
apscheduler>=3.10
playwright>=1.40
ollama>=0.1.7
tabulate>=0.9
dnspython>=2.4
flask>=3.0
ddgs>=9.0
gunicorn>=21.2
